moved StateT, state
