+++ modules/docs/arrow-docs/docs/docs/optics/getter/README.md
@@ -47,7 +47,7 @@ val tupleGetter: Getter<Tuple2<String, Int>, String> = Tuple2.first<String, Int>
 `Getter` also has some convenience methods to make working with [Reader]({{ '/docs/arrow/data/reader' | relative_url }}) easier.
 
 ```kotlin:ank
-import arrow.data.*
+import arrow.mtl.*
 
 val reader: Reader<NonEmptyList<String>, String> = NonEmptyList.head<String>().asGetter().ask()
