+++ modules/effects/arrow-effects-extensions/src/main/kotlin/arrow/effects/extensions/statet.kt
@@ -5,12 +5,12 @@ import arrow.core.Option
 import arrow.core.Some
 import arrow.core.Tuple2
 import arrow.core.getOrElse
-import arrow.data.StateT
-import arrow.data.StateTOf
-import arrow.data.StateTPartialOf
-import arrow.data.extensions.StateTMonadThrow
-import arrow.data.fix
-import arrow.data.runM
+import arrow.mtl.StateT
+import arrow.mtl.StateTOf
+import arrow.mtl.StateTPartialOf
+import arrow.mtl.extensions.StateTMonadThrow
+import arrow.mtl.fix
+import arrow.mtl.runM
 import arrow.effects.Ref
 import arrow.effects.typeclasses.Async
 import arrow.effects.typeclasses.Bracket
