+++ modules/mtl/arrow-mtl-data/src/main/kotlin/arrow/mtl/State.kt
@@ -1,4 +1,4 @@
-package arrow.data
+package arrow.mtl
 
 import arrow.core.ForId
 import arrow.core.Id
