+++ modules/mtl/arrow-mtl-data/src/test/kotlin/arrow/mtl/StateTests.kt
@@ -1,4 +1,4 @@
-package arrow.data
+package arrow.mtl
 
 import arrow.core.Id
 import arrow.core.Tuple2
