+++ modules/optics/arrow-optics/src/main/kotlin/arrow/optics/Getter.kt
@@ -9,8 +9,7 @@ import arrow.core.compose
 import arrow.core.identity
 import arrow.core.toT
 import arrow.mtl.Reader
-import arrow.data.State
-import arrow.data.map
+import arrow.mtl.State
 import arrow.mtl.map
 import arrow.higherkind
 import arrow.typeclasses.Monoid
