+++ modules/optics/arrow-optics/src/test/kotlin/arrow/optics/GetterTest.kt
@@ -6,10 +6,10 @@ import arrow.core.Some
 import arrow.core.Tuple2
 import arrow.core.extensions.monoid
 import arrow.core.toT
-import arrow.data.State
+import arrow.mtl.State
 import arrow.core.k
-import arrow.data.map
-import arrow.data.run
+import arrow.mtl.map
+import arrow.mtl.run
 import arrow.mtl.runId
 import arrow.test.UnitSpec
 import arrow.test.generators.functionAToB
