+++ modules/optics/arrow-optics/src/test/kotlin/arrow/optics/SetterTest.kt
@@ -8,8 +8,8 @@ import arrow.core.right
 import arrow.core.toT
 import arrow.core.extensions.option.functor.functor
 import arrow.core.getOrElse
-import arrow.data.State
-import arrow.data.run
+import arrow.mtl.State
+import arrow.mtl.run
 import arrow.test.UnitSpec
 import arrow.test.generators.functionAToB
 import arrow.test.generators.option
